ADS/requirements.txt
Flask
PyMySQL
DBUtils
fpdf2
bcrypt
python-dotenv
